compressed the images so that they load faster on the gallery page
